requirements.txt
Flask==3.0.2
boto3==1.34.79
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1